Split runtime restart and background orchestration
